Fix the capacity of poolable for Vec<T>
A vector can allocate more space than the current number of elements,
this allows it to cheaply grow the elements.  This is what
vec.capacity() returns.  vec.len() returns the actual number of items
in use, this is also the number of items you will get when slicing
with e.g. vec[..].  This latter matters since this is how the block is
meant to be used: a lot of consumers silce the block.  Hence vec.len()
gives the actual capacity of the block.  This is also reflected with
how the block (re)allocates capacity: vec.resize() changes the Vec's
len with no guarantees how it affects the Vec capacity.